New scripts folder
Added a scripts folder which includes the scripts that export functions for Vue components.
